+++ apps/client/src/pages/onBoarding/hooks/useOnboardingFunnel.ts
@@ -1,0 +1,157 @@
+import { sendGAEvent } from '@pinback/design-system/ui';
+import { usePostSignUp } from '@shared/apis/queries';
+import { useFunnel } from '@shared/hooks/useFunnel';
+import { useCallback, useEffect, useState } from 'react';
+import { AlarmsType } from '@constants/alarms';
+import { normalizeTime } from '@pages/onBoarding/utils/formatRemindTime';
+import { registerServiceWorker } from '@pages/onBoarding/utils/registerServiceWorker';
+import { Step, stepOrder, StepType } from '@pages/onBoarding/constants/onboardingSteps';
+import { firebaseConfig } from '../../../firebase-config';
+import { getApp, getApps, initializeApp } from 'firebase/app';
+import { getMessaging, getToken } from 'firebase/messaging';
+
+type AlarmSelection = 1 | 2 | 3;
+
+export function useOnboardingFunnel() {
+  const { mutate: postSignData } = usePostSignUp();
+  const { currentStep: step, currentIndex, setStep, goNext, goPrev } =
+    useFunnel<StepType>({
+      steps: stepOrder,
+      initialStep: Step.STORY_0,
+    });
+
+  const [direction, setDirection] = useState(0);
+  const [alarmSelected, setAlarmSelected] = useState<AlarmSelection>(1);
+  const [isMac, setIsMac] = useState(false);
+  const [userEmail, setUserEmail] = useState('');
+  const [remindTime, setRemindTime] = useState('09:00');
+  const [fcmToken, setFcmToken] = useState<string | null>(null);
+  const [jobShareAgree, setJobShareAgree] = useState(true);
+
+  useEffect(() => {
+    const storedEmail = localStorage.getItem('email');
+    if (storedEmail) {
+      setUserEmail(storedEmail);
+    }
+  }, []);
+
+  const requestFCMToken = useCallback(async (): Promise<string | null> => {
+    try {
+      const app = getApps().length > 0 ? getApp() : initializeApp(firebaseConfig);
+      const messaging = getMessaging(app);
+      const permission = await Notification.requestPermission();
+      registerServiceWorker();
+
+      if (permission !== 'granted') {
+        alert('알림 권한 허용이 필요합니다!');
+        return null;
+      }
+
+      const forFcmtoken = await getToken(messaging, {
+        vapidKey: import.meta.env.VITE_FIREBASE_VAPID_KEY,
+      });
+
+      if (forFcmtoken) {
+        return forFcmtoken;
+      }
+
+      alert('토큰 생성 실패. 다시 시도해주세요.');
+      return null;
+    } catch {
+      alert('알림 설정 중 오류가 발생했습니다. 다시 시도해주세요.');
+      return null;
+    }
+  }, []);
+
+  useEffect(() => {
+    const ua = navigator.userAgent.toLowerCase();
+    if (ua.includes('mac os') || ua.includes('iphone') || ua.includes('ipad')) {
+      setIsMac(true);
+    }
+
+    (async () => {
+      const token = await requestFCMToken();
+      if (token) {
+        setFcmToken(token);
+        localStorage.setItem('FcmToken', token);
+      } else {
+        alert('푸시 알람 설정 에러');
+      }
+    })();
+  }, [requestFCMToken]);
+
+  const nextStep = useCallback(async () => {
+    const next = stepOrder[currentIndex + 1];
+    const isAlarmStep = step === Step.ALARM;
+    const isFinalStep = step === Step.FINAL;
+    const isMacStep = next === Step.MAC;
+    const shouldSkipMacStep = isMacStep && !isMac;
+
+    if (isAlarmStep) {
+      if (alarmSelected === 1) setRemindTime('09:00');
+      else if (alarmSelected === 2) setRemindTime('20:00');
+      else {
+        const raw = AlarmsType[alarmSelected - 1].time;
+        setRemindTime(normalizeTime(raw));
+      }
+    }
+
+    if (shouldSkipMacStep) {
+      setDirection(1);
+      setStep(Step.FINAL);
+      return;
+    }
+
+    if (isFinalStep) {
+      postSignData(
+        { email: userEmail, remindDefault: remindTime, fcmToken },
+        {
+          onSuccess: () => (window.location.href = '/'),
+          onError: () => {
+            const savedEmail = localStorage.getItem('email');
+            if (savedEmail) window.location.href = '/';
+          },
+        }
+      );
+      return;
+    }
+
+    setDirection(1);
+    goNext();
+    sendGAEvent(
+      `onboard-step-${currentIndex + 1}`,
+      `onboard-step-${currentIndex + 1}`,
+      `onboard-step-${currentIndex + 1}`
+    );
+  }, [
+    alarmSelected,
+    currentIndex,
+    fcmToken,
+    goNext,
+    isMac,
+    postSignData,
+    remindTime,
+    setStep,
+    step,
+    userEmail,
+  ]);
+
+  const prevStep = useCallback(() => {
+    if (currentIndex > 0) {
+      setDirection(-1);
+      goPrev();
+    }
+  }, [currentIndex, goPrev]);
+
+  return {
+    step,
+    currentIndex,
+    direction,
+    alarmSelected,
+    jobShareAgree,
+    setAlarmSelected,
+    setJobShareAgree,
+    nextStep,
+    prevStep,
+  };
+}
